% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_annots_to_nodes}
\alias{add_annots_to_nodes}
\title{Add annotations to nodes}
\usage{
add_annots_to_nodes(ccog, node_annots, id_column = "")
}
\arguments{
\item{ccog}{CCC graph}

\item{node_annots}{annotation table (data.frame).  It should include
an ID column which should be used in left_join.}

\item{id_column}{ID column to be used for join operation between tables}
}
\value{
CCC graph with revised annotations for nodes
}
\description{
Add annotations to nodes
}
