% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_node_annot_to_edges}
\alias{add_node_annot_to_edges}
\title{Add node annotations to edges}
\usage{
add_node_annot_to_edges(ccog, node_annot_colname)
}
\arguments{
\item{ccog}{CCC graph}

\item{node_annot_colname}{column names for annotations to be added to edges}
}
\value{
CCC graph with revised annotations for edges
}
\description{
Annotations are added to 'src' and 'dst' nodes, and
those are suffixed with "_src" and "_dst", respectively.
}
