% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_spatial_ccc_graph_metrics_to_edges}
\alias{add_spatial_ccc_graph_metrics_to_edges}
\title{Copy graph metrics from nodes to edges}
\usage{
add_spatial_ccc_graph_metrics_to_edges(sp_ccc_graph, from_scratch = TRUE)
}
\description{
Internal function
}
