% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_spatial_coords}
\alias{add_spatial_coords}
\title{Add spatial coordinates to CCC table}
\usage{
add_spatial_coords(df, spatial_coords)
}
\arguments{
\item{df}{CCC table. see \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}.}

\item{spatial_coords}{spatial coordinates: cell_id, spot_x, spot_y;
the name does not matter, but only the order as they will be renamed.}
}
\description{
internal function
}
