% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_spatial_data}
\alias{add_spatial_data}
\title{Add spatial data to CCC table}
\usage{
add_spatial_data(df, spatial_data)
}
\arguments{
\item{df}{CCC table. see \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}.}

\item{spatial_data}{spatial data: cell_id, spot_x, spot_y, ...;
they should contain cell_id, spot_x, spot_y, and the rest will
be added}
}
\description{
internal function
}
