% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{amend_ccc_table_with_cell_annots}
\alias{amend_ccc_table_with_cell_annots}
\title{Amend CCC table with annotations for source and destination cells}
\usage{
amend_ccc_table_with_cell_annots(ccc_table, spe, annot_cols, overwrite = TRUE)
}
\arguments{
\item{ccc_table}{...}

\item{spe}{...}

\item{annot_cols}{...}
}
\description{
Amend CCC table with annotations for source and destination cells
}
