% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{barebone_spatial_ccc_graph}
\alias{barebone_spatial_ccc_graph}
\title{Create barebone CCC graph}
\usage{
barebone_spatial_ccc_graph(spe, spot_dist_cutoff = 1.5)
}
\arguments{
\item{spe}{SpatialExperiment object}
}
\value{
spatial CCC graph w/o LRscore
}
\description{
Create barebone CCC graph
}
