% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{calc_spot_dist0}
\alias{calc_spot_dist0}
\title{Calculate distance between spots}
\usage{
calc_spot_dist0(coords, long_format = TRUE)
}
\arguments{
\item{coords}{spatial coordinates in data.frame
with \code{cell_id}, \code{x}, and \code{y} in that order.  The column headings
do not matter, just the order.}

\item{long_format}{if TRUE (default), return long format data.frame.
Otherwise, matrix format.}
}
\value{
A data fame with computed distance including both \code{d} and \code{norm.d}.
\code{norm.d} is a multiple of the shortest distance between spots.
}
\description{
Calculate distance between spots
}
