% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{calculate_percent_rank}
\alias{calculate_percent_rank}
\title{Calculate (combined) percent rank of n and LRscore}
\usage{
calculate_percent_rank(ccc_graph_ls)
}
\arguments{
\item{ccc_graph_ls}{list of spatial CCC graph}
}
\value{
data frame (tibble)

In the table returned,
\itemize{
\item n:  number of edges in CCC graph (LR),
\item LRscore: median of all LRscores in CCC graph
\item n_perc_rank:  percent rank of n
\item LRscore_perc_rank: percent rank of LRscore (median)
\item perc_rank: combined percent rank (geometric mean of n_perc_rank and LRscore_perc_rank)
\item perc_x: binned perc_rank
}
}
\description{
Calculate (combined) percent rank of n and LRscore
}
