% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_op.R
\name{ccc_graph_intersect}
\alias{ccc_graph_intersect}
\title{Create a new spatial CCC graph with common edges}
\usage{
ccc_graph_intersect(ccog1, ccog2)
}
\arguments{
\item{ccog1, ccog2}{spatial CCC graphs}
}
\value{
a spatial CCC graph with common edges
}
\description{
Create a new spatial CCC graph with common edges
}
