% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_op.R
\name{ccc_graph_subtract}
\alias{ccc_graph_subtract}
\title{Subtract a spatial CCC graph (ccog2) from another (ccog1)}
\usage{
ccc_graph_subtract(ccog1, ccog2)
}
\arguments{
\item{ccog1, ccog2}{spatial CCC graphs}
}
\value{
a spatial CCC graph with common edges
}
\description{
Subtract a spatial CCC graph (ccog2) from another (ccog1)
}
