% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_op.R
\name{ccc_graph_union}
\alias{ccc_graph_union}
\title{Combine two CCC graphs}
\usage{
ccc_graph_union(ccog1, ccog2)
}
\arguments{
\item{ccog1, ccog2}{spatial CCC graphs}
}
\value{
flattened spatial CCC graph
}
\description{
Combine two CCC graphs
}
