% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_across_clusters.R
\name{count_interactions_across_clusters_ccc_graph_tbl}
\alias{count_interactions_across_clusters_ccc_graph_tbl}
\alias{summarize_ccc_graph_tbl}
\alias{count_interactions_across_clusters_ccc_graph}
\alias{summarize_ccc_graph}
\title{Count interactions across clusters}
\usage{
count_interactions_across_clusters_ccc_graph_tbl(cog_tbl, include_edges = TRUE)

summarize_ccc_graph_tbl(cog_tbl, include_edges = TRUE)

count_interactions_across_clusters_ccc_graph(cog, include_edges = TRUE)

summarize_ccc_graph(cog, include_edges = TRUE)
}
\arguments{
\item{cog_tbl}{table (data.frame) converted from ccc_graph_list.
Node and edge statistics are calculated for each
(LR, cluster_src, cluster_dst) tuple.}

\item{include_edges}{If TRUE (default), include edge counts.}

\item{cog}{spatial CCC graph}
}
\value{
data frame with the counts of nodes and edges interacting
across clusters
}
\description{
Count interactions across clusters
}
