% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_distance.R
\name{dist_cell_overlap_ccc_tbl}
\alias{dist_cell_overlap_ccc_tbl}
\title{Compute cell overlap distance}
\usage{
dist_cell_overlap_ccc_tbl(ccc_tbl)
}
\arguments{
\item{ccc_tbl}{CCC graph in tabular format}
}
\value{
Jaccard distance in matrix format
}
\description{
Compute cell overlap distance
}
