% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_distance.R
\name{dist_cell_overlap_edgesets}
\alias{dist_cell_overlap_edgesets}
\alias{lf_cell_overlap_edgesets}
\title{Compute distance between LRs}
\usage{
dist_cell_overlap_edgesets(edgeset_list)

lf_cell_overlap_edgesets(edgeset_list)
}
\arguments{
\item{edgeset_list}{a set of concatenated edges
(see \code{\link[=prep_cell_overlap_ccc_tbl]{prep_cell_overlap_ccc_tbl()}})}
}
\description{
\code{dist_cell_overlap_edgesets()} returns Jaccard distance matrix.
\code{lf_cell_overlap_edgesets()} returns Jaccard distance in long format.

Calculate Jaccard distance between LRs where each LR has
a set of concatenated edges
}
