% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{extract_ccc_graph_metrics}
\alias{extract_ccc_graph_metrics}
\title{Extract spatial CCC graph metrics}
\usage{
extract_ccc_graph_metrics(ccc_graph, level = c("graph", "group"))
}
\arguments{
\item{ccc_graph}{an output by \code{\link[=to_spatial_ccc_graph]{to_spatial_ccc_graph()}}}

\item{level}{extract graph metrics from either overall graph ("graph") or
subgraph ("group")}
}
\value{
one row data.frame with graph metrics
}
\description{
Extract spatial CCC graph metrics
}
