% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{extract_spatial_ccc_graph_edges}
\alias{extract_spatial_ccc_graph_edges}
\title{Convert spatial CCC graphs to spatial CCC table}
\usage{
extract_spatial_ccc_graph_edges(ccc_graph_ls)
}
\arguments{
\item{ccc_graph_ls}{a list of spatial CCC graphs}
}
\value{
spatial CCC table
}
\description{
Convert spatial CCC graphs to spatial CCC table
}
