% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{geom_spatial}
\alias{geom_spatial}
\title{Geom Spatial to visualize tissue image}
\usage{
geom_spatial(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\description{
internal function
}
