% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LRdb.R
\name{get_LRdb}
\alias{get_LRdb}
\title{Return LRdb table}
\usage{
get_LRdb(species = c("human", "mouse"))
}
\arguments{
\item{species}{string, either "human" (default) or "mouse"}
}
\value{
LRdb table in data frame
}
\description{
Return LRdb table
}
\examples{
LRdb <- get_LRdb(species = "human")

}
