% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{get_spatial_data}
\alias{get_spatial_data}
\title{Get spatial coordinates}
\usage{
get_spatial_data(spe)
}
\arguments{
\item{spe}{SpatialExperiment object}
}
\value{
spatial coordinates in data.frame with
\code{cell_id}, ... (from \code{spatialCoords(spe)}),
\code{spot_x}, \code{spot_y}.
}
\description{
Returns default spatial coordinates and scaled coordinates,
\code{spot_x} and \code{spot_y}.
}
