% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{interaction_strength}
\alias{interaction_strength}
\title{Calculate interaction strength}
\usage{
interaction_strength(x, eps = 1/(1.73 * 2))
}
\arguments{
\item{x}{distance}

\item{eps}{scaling factor for calculating diffusion/interaction strength.
default value = 1/(1.73*2) corresponds that the spots with ~200nm retain
~50\% of cell signal,}
}
\value{
interaction strength
}
\description{
Calculate interaction strength
}
