% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_distance.R
\name{lf_to_dist}
\alias{lf_to_dist}
\title{Convert distance in long format to matrix}
\usage{
lf_to_dist(lf)
}
\arguments{
\item{lf}{distance in long format}
}
\value{
distance in matrix format
}
\description{
It ensures columns and rows have the same features
}
