% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRdb_mouse}
\alias{LRdb_mouse}
\title{Ligand-receptor database (mouse)}
\format{
\subsection{\code{LRdb_mouse}}{

A data frame with 2,033 rows and 10 columns:
\describe{
\item{LR}{ligand_receptor pair}
\item{ligand_gene_symbol, receptor_gene_symbol}{ligand & receptor gene symbols}
\item{ligand_gene_id, receptor_gene_id}{ligand & receptor gene ids - Entrez ID}
...
}
}
}
\source{
\url{http://tcm.zju.edu.cn/celltalkdb/download.php}
}
\usage{
LRdb_mouse
}
\description{
ligand-receptor database for mouse
}
\keyword{datasets}
