% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_across_clusters.R
\name{map_ccc_graph_across_clusters}
\alias{map_ccc_graph_across_clusters}
\alias{map_ccc_graph_across_clusters_shortcut}
\title{Cross-reference ccc_graph across clusters}
\usage{
map_ccc_graph_across_clusters(cog_tbl, sp_dist)

map_ccc_graph_across_clusters_shortcut(cog_tbl, sp_dist)
}
\arguments{
\item{cog_tbl}{table (data.frame) converted from ccc_graph_list.
Node and edge statistics are calculated for each
(LR, cluster_src, cluster_dst) tuple.}

\item{sp_dist}{an output of calc_spot_dist() amended by cluster IDs
(cluster_src, cluster_dst) should be used for cluster IDs.
The easiest way to make sure this is satisfied is to add
"cluster" column in SpatialExperiment object with cluster/group
information.  This will be automatically used in
calc_spot_dist() function when it creates spot_dist data frame.}
}
\value{
data frame with estimated p-values
}
\description{
Cross-reference \emph{ccc_graph} across clusters defined for each spot and
test the frequency of inter-cluster communications and
calculate the likelihood of the frequency by chance.
}
\details{
ccc_graph is summarized in \emph{cog_tbl} and
spatial configuration is stored in \emph{sp_dist} which
also includes (cluster_src, cluster_dst).

\strong{Step 1} Create base distribution by cross-referencing
\emph{sp_dist} against cluster configuration

\strong{Step 2} Map ccc_graph over clusters and take the average across
LRs to create "null" distribution of inter/intra-cluster
interaction frequency.

\strong{Step 3} Calculate the likelihood of observed inter/intra-cluster
interaction frequency for each LR by chance using the "null"
distribution estimated in \strong{Step 2}.

\emph{map_ccc_graph_across_clusters_shortcut()} maps and summarize
the counts by LR only, not by (LR, cluster_src, cluster_dst) like
map_ccc_graph_across_clusters().
}
