% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{modify_alpha_image}
\alias{modify_alpha_image}
\title{Modify image with alpha}
\usage{
modify_alpha_image(raster, image.alpha = NA)
}
\arguments{
\item{raster}{Raster image, like imgRaster(spe)}

\item{image.alpha}{alpha value for tissue_img. If NA, keep the current value.}
}
\description{
Modify image with alpha
}
