% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{normalize_for_LR_specificity}
\alias{normalize_for_LR_specificity}
\alias{normalize_gexp_for_LR_specificity}
\title{Normalize gene expression data}
\usage{
normalize_for_LR_specificity(spe, assay_name = "logcounts")

normalize_gexp_for_LR_specificity(gexp)
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{assay_name}{assay name in string}

\item{gexp}{gene expression matrix}
}
\value{
SpatialExperiment object
in \code{normalize_for_LR_specificity(...)}

TF-IDF normalized gene expression matrix
in \code{normalize_gexp_for_LR_specificity(...)}
}
\description{
Normalize gene expression data via
term frequency-inverse document frequency (TF-IDF) transformation
to maximize LR specificity.
}
\details{
Gene expression itself cannot represent the specificity of
ligand/receptor binding. Zhang et al. proposed
term frequency-inverse document frequency (TF-IDF) transformation
to improve the specificity of ligand/receptor binding.

\deqn{E = e_{i,j} / \sum_{i=j}^m e_{i,j} \cdot \log(1 + n/e_{i,j}) \cdot m}

where \eqn{e_{i,j}} is the expression value of gene \eqn{i} in cell \eqn{j},
with \eqn{m} genes and \eqn{n} cells.

\strong{Ref:} 2.4 Pipeline for intercellular communication analysis
In \href{https://doi.org/10.1093/bioinformatics/btab036}{Zhang et al., Bioinformatics 37(14): 2025-2032}
}
