% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spatial_feature}
\alias{plot_spatial_feature}
\title{Plot spatial feature}
\usage{
plot_spatial_feature(
  spe,
  feature,
  cells_of_interest = NULL,
  spot_size = 1.6,
  spot_alpha = 1,
  image_alpha = NA,
  show_tissue_image = TRUE,
  clip = TRUE
)
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{feature}{string, column name for a feature to be visualized}

\item{cells_of_interest}{array of cell ids to highlight.}

\item{spot_size}{specify size of spot (feature).  Currently,
default values do not work well.  Might need some manual adjustments.}

\item{spot_alpha}{specify transparency of spot (feature)}

\item{image_alpha}{alpha value for tissue image.
If NA, keep the current value.}

\item{show_tissue_image}{if TRUE, tissue image is shown.}

\item{clip}{if TRUE (dafault), tissue image will be cliped to the size to
cover only the spots}
}
\description{
Visualize spatial features
w/ or w/o tissue image
}
\details{
The spots are placed on their original locations,
keeping spatial relations.
}
\examples{
\dontrun{
plot_spatial_features(spe,
                      feature = "mclust_6")

plot_spatial_features(spe,
                      feature = "mclust_6",
                      image_alpha = 0.75)
}

}
