% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_distance.R
\name{prep_cell_overlap_ccc_tbl}
\alias{prep_cell_overlap_ccc_tbl}
\title{Prepare CCC table for distance calcuation}
\usage{
prep_cell_overlap_ccc_tbl(ccc_tbl)
}
\arguments{
\item{ccc_tbl}{CCC table}
}
\value{
a list of CCC graph which is a list of concatenated edges
}
\description{
Reorganize CCC table into a list where each element is a CCC graph,
an edge list where each edge, (src, dst) pair, is not concatenated
into src_dst string for easy enumeration later.
}
