% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{set_ccc_table_default_cluster}
\alias{set_ccc_table_default_cluster}
\title{Set default cluster in CCC table}
\usage{
set_ccc_table_default_cluster(ccc_table, cluster_name)
}
\arguments{
\item{ccc_table}{CCC table, output of \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}}

\item{cluster_name}{string, the name of default cluster.
{cluster_name}.src and {cluster_name}.dst columns will be copied to
"cluster.src" and "cluster.dst".}
}
\value{
CCC table with amended columns
}
\description{
Set default cluster in CCC table
}
