% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialCCC-package.R
\docType{package}
\name{spatialCCC-package}
\alias{spatialCCC}
\alias{spatialCCC-package}
\title{spatialCCC: Spatial Transcriptomic cell-cell communication analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Investigate cell-cell signaling, by analyzing ligand-receptor interactions in spatial transcriptomic data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dolchan/spatialCCC}
  \item Report bugs at \url{https://github.com/dolchan/spatialCCC/issues}
}

}
\author{
\strong{Maintainer}: Seungchan Kim \email{dolchan@gmail.com} (\href{https://orcid.org/0000-0003-4183-4338}{ORCID})

}
\keyword{internal}
