% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{spatialCCC_example}
\alias{spatialCCC_example}
\title{List example files for spatialCCC package}
\usage{
spatialCCC_example(path = NULL)
}
\arguments{
\item{path}{path to an example file}
}
\value{
a list of example files if called with no parameter, or
the absolute path to an example file if called with an example file
}
\description{
List example files for spatialCCC package
}
