% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{split_ccc_tbl_LR}
\alias{split_ccc_tbl_LR}
\title{Split LR into L and R}
\usage{
split_ccc_tbl_LR(ccc_tbl)
}
\arguments{
\item{ccc_tbl}{an output of \code{\link[=extract_spatial_ccc_graph_edges]{extract_spatial_ccc_graph_edges()}} or
any table with "LR" column which should be in "$L_$R" format}
}
\description{
Split LR into L and R
}
