% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{summarise_ccc_by_cell}
\alias{summarise_ccc_by_cell}
\title{Summarize CCC table by (cell, LR)}
\usage{
summarise_ccc_by_cell(ccc_tbl)
}
\arguments{
\item{ccc_tbl}{CCC graph table.  see \code{\link[=compute_spatial_ccc_tbl]{compute_spatial_ccc_tbl()}}.}
}
\description{
internal function
}
