% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_graph_across_clusters.R
\name{summarize_by_LR_pval_across_clusters}
\alias{summarize_by_LR_pval_across_clusters}
\title{Summarize p-value across clusters}
\usage{
summarize_by_LR_pval_across_clusters(LR_pval_across)
}
\arguments{
\item{LR_pval_across}{data.frame, an output of map_ccc_graph_across_clusters.
the columns include cluster_src, cluster_dst, n_nodes, p_val_node,
n_edges, p_val_edge}
}
\description{
Summarize p-value (LR, cluster_src, cluster_dst)
across (cluster_src, cluster_dst).
}
