% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{summarize_ccc_graph_metrics}
\alias{summarize_ccc_graph_metrics}
\title{Summarize spatial CCC graph list to table}
\usage{
summarize_ccc_graph_metrics(ccc_graph_ls, level = c("graph", "group"))
}
\arguments{
\item{ccc_graph_ls}{list of spatial CCC graph,
each of which is an output of to_spatial_ccc_graph.}

\item{level}{extract graph metrics from either overall graph ("graph") or
subgraph ("group")}
}
\value{
graph metrics table summarized for each LR pair
}
\description{
Summarize spatial CCC graph list to table
}
