% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{tag_cells_in_ccc_graph}
\alias{tag_cells_in_ccc_graph}
\title{Tag nodes if included in COIs}
\usage{
tag_cells_in_ccc_graph(ccog, COIs, edges_expanded_to_group = FALSE)
}
\arguments{
\item{ccog}{CCC graph}

\item{COIs}{array of nodes to highlight}

\item{edges_expanded_to_group}{If TRUE,}
}
\value{
CCC graph with InCOI and tagged columns (see above)
}
\description{
Add additional columns to all nodes in CCC graph:
\enumerate{
\item InCOI = TRUE if a node is included in COIs or is connected
to a node in COIs
\item tagged = TRUE if a node is included in a group that includes
any of COIs, when edges_expanded_to_group = TRUE
}
}
