% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{tidy_up_ccc_graph}
\alias{tidy_up_ccc_graph}
\title{Tidy up CCC graph by removing isolated nodes}
\usage{
tidy_up_ccc_graph(ccog)
}
\arguments{
\item{ccog}{CCC graph}
}
\value{
CCC graph with all isolated nodes removed.
}
\description{
Tidy up CCC graph by removing isolated nodes
}
