% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_barebone_spatial_ccc_graph}
\alias{to_barebone_spatial_ccc_graph}
\title{Convert CCC table to barebone spatial CCC graph}
\usage{
to_barebone_spatial_ccc_graph(ccc_table)
}
\arguments{
\item{ccc_table}{an output of \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}}
}
\value{
CCC graph (barebone)
}
\description{
Convert CCC table to barebone spatial CCC graph
}
