% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_LR_summary_tbl}
\alias{to_LR_summary_tbl}
\title{Summarize spatial CCC graph for LR}
\usage{
to_LR_summary_tbl(ccc_graph_ls)
}
\arguments{
\item{ccc_graph_ls}{list of spatial CCC graph,
each of which is an output of to_spatial_ccc_graph.}
}
\value{
spatial CCC graph summarized for LR,
aggregating LRscores over the graph and computing percentile ranks.
}
\description{
Summarize spatial CCC graph for LR
}
\section{CCC_graph_percent_rank}{


In the table returned,
\itemize{
\item n:  number of edges in CCC graph (LR),
\item LRscore: median of all LRscores in CCC graph
\item n_perc_rank:  percent rank of n
\item LRscore_perc_rank: percent rank of LRscore (median)
\item perc_rank: combined percent rank (geometric mean of n_perc_rank and LRscore_perc_rank)
\item perc_group: binned perc_rank
}
}

