% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_spatial_ccc_graph}
\alias{to_spatial_ccc_graph}
\title{Convert CCC table to spatial CCC graph}
\usage{
to_spatial_ccc_graph(ccc_table, spatial_data, LR_of_interest = NULL)
}
\arguments{
\item{ccc_table}{an output of \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}}

\item{spatial_data}{a table with spatial coordinates of
spots/cells in spatial transcriptomic data.  if spatial_data
includes cell_id, spot_x, spot_y, they will be used as is;
Otherwise, use the first three columns as if they are
cell_id, spot_x, spot_y}

\item{LR_of_interest}{LR pair name in string.  If NULL (default),
CCC graph will be created using all LR pairs.}
}
\description{
Convert CCC table to spatial CCC graph
}
