% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_spatial_ccc_tbl}
\alias{to_spatial_ccc_tbl}
\title{Convert CCC graph to CCC table}
\usage{
to_spatial_ccc_tbl(ccc_graph)
}
\arguments{
\item{ccc_graph}{an output of \code{\link[=compute_spatial_ccc_graph]{compute_spatial_ccc_graph()}}}
}
\value{
spatial CCC table -- \code{tibble}
}
\description{
Convert CCC graph to CCC table
}
