% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{transfer_node_annots_to_edges}
\alias{transfer_node_annots_to_edges}
\title{Add node annotations to edges}
\usage{
transfer_node_annots_to_edges(
  ccog,
  node_annot_colname,
  target_name = node_annot_colname
)
}
\arguments{
\item{ccog}{CCC graph}

\item{node_annot_colname}{column names for annotations to be added to edges}

\item{target_name}{if specified, node_annot_colname will be renamed to target_name}
}
\value{
CCC graph with revised annotations for edges.
Returns null If CCC graph is null.
}
\description{
Note annotations are transferred to edges, and
those are suffixed with "_src" and "_dst", respectively.
}
