% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{add_spatial_ccc_graph_metrics_to_edges}
\alias{add_spatial_ccc_graph_metrics_to_edges}
\title{Copy graph metrics from nodes to edges}
\usage{
add_spatial_ccc_graph_metrics_to_edges(sp_ccc_graph, from_scratch = TRUE)
}
\arguments{
\item{sp_ccc_graph}{an output of \code{\link[=to_spatial_ccc_graph]{to_spatial_ccc_graph()}}}

\item{from_scratch}{if TRUE, the existing metrics are wiped clean.}
}
\description{
Internal function
}
