% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{calc_spot_dist}
\alias{calc_spot_dist}
\title{Calculate distance between spots}
\usage{
calc_spot_dist(spe, long_format = TRUE, spot_dist_cutoff = 1.5)
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{long_format}{if TRUE (default), return long format data.frame.
Otherwise, matrix format.}

\item{spot_dist_cutoff}{cutoff value for norm.d in spot
distances. Default cutoff is 1.5 and
it comes from sqrt(3) ~ 1.73, based on how Visium spot-array is
arranged, so by default, only computing nearest neighbors.}
}
\value{
A data fame with computed distance including both \code{d} and \code{norm.d}.
\code{norm.d} is a multiple of the shortest distance between spots.
}
\description{
Calculate distance between spots
}
