% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{calc_spot_dist}
\alias{calc_spot_dist}
\title{Calculate distance between spots}
\usage{
calc_spot_dist(
  spe,
  long_format = TRUE,
  include_cluster = TRUE,
  spot_dist_cutoff = 1.5
)
}
\arguments{
\item{spe}{SpatialExperiment object}

\item{long_format}{if TRUE (default), return long format data.frame.
Otherwise, matrix format.}

\item{include_cluster}{if TRUE, include \emph{cluster_src} and \emph{cluster_dst}
in the output data frame.}

\item{spot_dist_cutoff}{maximum distance allowed between spots}
}
\value{
A data fame with computed distance including both \code{d} and \code{norm.d}.
\code{norm.d} is a multiple of the shortest distance between spots.

The data frame will only contain the spots with the distance (\code{norm.d})
less than \code{spot_dist_cutoff}.  The default cutoff (\code{spot_dist_cutoff} = 1.5)
will leave only those with immediately neighboring cells.
}
\description{
Calculate distance between spots
}
