% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{flatten_ccc_graph_list}
\alias{flatten_ccc_graph_list}
\title{Flatten multiple CCC graphs into aggregate one}
\usage{
flatten_ccc_graph_list(ccc_graph_ls)
}
\arguments{
\item{ccc_graph_ls}{a list of spatial CCC graphs}
}
\value{
a spatial CCC graph
}
\description{
Flatten multiple CCC graphs into aggregate one
}
\section{flattening_CCC_graph}{


The following node parameters will be kept:
\itemize{
\item name
\item spot_x, spot_y
\item in_tissue
\item array_row, array_col
\item sizeFactor
}

The following node parameters will be aggregated over all LR pairs:
\itemize{
\item n, n.src, n.dst, LRscore.sum.src, and LRscore.sum.dst will be summed.
\item node_is_src and node_is_dst will be OR'd
\item n.inflow and LRscore.sum.inflow will be summed.
}
}

