% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_visualization.R
\name{plot_spatial_ccc_graph}
\alias{plot_spatial_ccc_graph}
\title{Plot spatial CCC graph}
\usage{
plot_spatial_ccc_graph(
  ccc_graph,
  graph_layout = "auto",
  cells_of_interest = NULL,
  edges_expanded_to_group = TRUE,
  edge_color = "group_n_edges",
  node_color = "group_n_nodes",
  edge_range = NULL,
  node_range = NULL,
  edge_width = 0.5,
  edge_alpha = 1,
  node_size = 0.5,
  node_alpha = 1,
  clip = FALSE,
  tissue_img = NULL,
  ghost_img = FALSE,
  which_on_top = "edge",
  show_arrow = FALSE
)
}
\arguments{
\item{ccc_graph}{spatial CCC graph for a LR pair}

\item{graph_layout}{string for graph layout algorithm. Default is "auto";
see Details for more information}

\item{cells_of_interest}{array of cell ids to highlight.}

\item{edges_expanded_to_group}{if TRUE and cells_of_interest is given,
the nodes and the edges are expanded to include all the nodes and edges that
belong to the groups that cells_of_interest belong to.}

\item{edge_color, node_color}{column name to be used
for edge and node colors, respectively}

\item{edge_range, node_range}{bound given in c(min, max) format.
if specified, the ranges are set using these values, instead of the ones
derived from data.  Useful when multiple plots are combined
later for comparison.}

\item{edge_width, node_size}{specify edge width and node size.  Currently,
default values do not work well.  Might need some manual adjustments.}

\item{edge_alpha, node_alpha}{specify transparency of edges and nodes}

\item{clip}{if TRUE, tissue_img will be cliped to show only active cell-cell
communications}

\item{tissue_img}{Raster image, like imgRaster(spe)}

\item{ghost_img}{if TRUE, tissue_img is not displayed, only used for
placing spots.}

\item{which_on_top}{either "edge" or "node"}

\item{show_arrow}{if TRUE, arrow is added to edge.  Default is FALSE.}
}
\description{
Visualize spatial CCC graph in various formats,
w/ or w/o tissue image
}
\details{
\itemize{
\item CCC graph with tissue image. In this format,
the spots are placed on their original locations, keeping
spatial relations.
\item CCC graph without tissue image. In this format,
the spots are placed using user-specified graph layout
algorithm.  Default is ("kk")\code{\link[igraph:layout_with_kk]{igraph::layout_with_kk()}} algorithm.
See \code{\link[ggraph:ggraph]{ggraph::create_layout()}} for available algorithms.
}
}
\examples{
\dontrun{
plot_spatial_ccc_graph(ccc_graph = ccc_graph_list[["App_Dcc"]],
                       tissue_img = imgRaster(spe),
                       node_color = "group_diameter",
                       node_size = 1,
                       edge_color = "group_diameter",
                       which_on_top = "edge")

plot_spatial_ccc_graph(ccc_graph = ccc_graph_list[["App_Dcc"]],
                       node_color = "group_diameter",
                       node_size = 1,
                       edge_color = "group_diameter",
                       which_on_top = "edge")

plot_spatial_ccc_graph(ccc_graph = ccc_graph_list[["App_Dcc"]],
                       graph_layout = "stress",
                       node_color = "group_diameter",
                       node_size = 1,
                       edge_color = "group_diameter",
                       which_on_top = "edge")
}

}
