% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{summarise_ccc_by_cell}
\alias{summarise_ccc_by_cell}
\title{Summarize CCC table by (cell, LR)}
\usage{
summarise_ccc_by_cell(ccc_tbl)
}
\arguments{
\item{ccc_tbl}{CCC graph table.  see \code{\link[=compute_spatial_ccc_tbl]{compute_spatial_ccc_tbl()}}.}
}
\description{
internal function
}
\details{
summarize LRscores (edge) to cell (node) by
\enumerate{
\item adding LRscores in all the edges connected to the cell as \code{src} or \code{dst}
LRscore.sum.src, LRscore.sum.dst
n.src, n.dst
\item n.inflow = n.dst - n.src
\item LRscore.sum.inflow = LRscore.sum.dst - LRscore.sum.src
}
}
