% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_spatial_ccc_graph}
\alias{to_spatial_ccc_graph}
\title{Convert CCC table to CCC graph}
\usage{
to_spatial_ccc_graph(ccc_tbl, spatial_data)
}
\arguments{
\item{ccc_tbl}{an output of \code{\link[=compute_spatial_ccc_tbl]{compute_spatial_ccc_tbl()}}}

\item{spatial_data}{a table with spatial coordinates of
spots/cells in spatial transcriptomic data.  if spatial_data
includes cell_id, spot_x, spot_y, they will be used as is;
Otherwise, use the first three columns as if they are
cell_id, spot_x, spot_y}
}
\value{
spatial CCC graph -- a \code{tidygraph}
}
\description{
Convert CCC table to CCC graph
}
\section{graph metrics}{

\enumerate{
\item For overall graph,
\itemize{
\item graph_n_nodes,
\item graph_n_edges
\item graph_component_count
\item graph_motif_count
\item graph_diameter
\item graph_un_diameter
\item graph_mean_dist
\item graph_circuit_rank = graph_n_edges - graph_n_nodes + graph_component_count
\item graph_reciprocity
\item graph_clique_num (sp_ccc_graph assumed as undirected)
\item graph_clique_count (sp_ccc_graph assumed as undirected)
}
\item For each sub-graph (after \code{\link[=group_components]{group_components()}})
\itemize{
\item group_n_nodes
\item group_n_edges
\item group_adhesion
\item group_motif_count
\item group_diameter
\item group_un_diameter
\item group_mean_dist
\item group_girth
\item group_circuit_rank
\item group_reciprocity
}
}
}

