% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_spatial_ccc_graph_list}
\alias{to_spatial_ccc_graph_list}
\title{Convert a list of CCC tables to a list of spatial CCC graphs}
\usage{
to_spatial_ccc_graph_list(ccc_tbl, spatial_coords, workers = 1)
}
\arguments{
\item{ccc_tbl}{an output of \code{\link[=compute_spatial_ccc]{compute_spatial_ccc()}}}

\item{spatial_coords}{a table with spatial coordinates of
spots/cells in spatial transcriptomic data.  if spatial_coords
includes cell_id, spot_x, spot_y, they will be used as is;
Otherwise, use the first three columns as if they are
cell_id, spot_x, spot_y}

\item{workers}{the number of processes to be used for parallel
processing}
}
\description{
Convert a list of CCC tables to a list of spatial CCC graphs
}
