% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ccc.R
\name{to_spatial_ccc_tbl}
\alias{to_spatial_ccc_tbl}
\title{Convert spatial CCC graphs to spatial CCC table}
\usage{
to_spatial_ccc_tbl(ccc_graph_list)
}
\arguments{
\item{ccc_graph_list}{a list of spatial CCC graphs}
}
\value{
spatial CCC table
}
\description{
Convert spatial CCC graphs to spatial CCC table
}
